fix: fix broken project groups permisions gf-569
