+++ apps/backend/src/modules/project-groups/libs/helpers/check-project-group-permission/check-project-group-permission.helper.ts
@@ -1,9 +1,8 @@
 import {
-	ExceptionMessage,
+	type PermissionKey,
 	type ProjectPermissionKey,
 } from "~/libs/enums/enums.js";
-import { checkHasPermission } from "~/libs/helpers/helpers.js";
-import { HTTPCode, HTTPError } from "~/libs/modules/http/http.js";
+import { checkUserPermissionsHelper } from "~/libs/helpers/helpers.js";
 import { type ValueOf } from "~/libs/types/types.js";
 import { type UserAuthResponseDto } from "~/modules/users/users.js";
 
@@ -12,26 +11,18 @@ import { type ProjectGroupGetAllItemResponseDto } from "../../types/types.js";
 const checkProjectGroupPermission = (payload: {
 	projectGroup: ProjectGroupGetAllItemResponseDto;
 	projectsPermissions: ValueOf<typeof ProjectPermissionKey>[];
+	rootPermissions: ValueOf<typeof PermissionKey>[];
 	user: UserAuthResponseDto;
 }): number => {
-	const { projectGroup, projectsPermissions, user } = payload;
+	const { projectGroup, projectsPermissions, rootPermissions, user } = payload;
 	const { id: projectId } = projectGroup.projectId;
 
-	const userProjectsPermissions = user.projectGroups
-		.filter((group) => projectId && group.projectId === projectId)
-		.flatMap((projectGroup) => projectGroup.permissions);
-
-	const hasProjectPermission = checkHasPermission(
+	checkUserPermissionsHelper({
+		projectId,
 		projectsPermissions,
-		userProjectsPermissions,
-	);
-
-	if (!hasProjectPermission) {
-		throw new HTTPError({
-			message: ExceptionMessage.NO_PERMISSION,
-			status: HTTPCode.FORBIDDEN,
-		});
-	}
+		rootPermissions,
+		user,
+	});
 
 	return projectGroup.id;
 };
